requirements.txt
python-telegram-bot==20.3
Flask==3.0.0
waitress==2.1.2
requests==2.31.0
pytz==2024.1
